Add files via upload
This is an assignment was given in by Ekemini Otu, the organiser of Artvion Squid Edition Programming.  In this assignment, you will search for words based on predefined criteria. For instance, which words does not include any vowels or the letters of which words are in alphabetical order...The words are from Lorem Ipsum. Lorem Ipsum is simply dummy text of the printing and typesetting industry.
(https://www.lipsum.com/)